Add operations

+++ trying-big-data/src/main/java/com/data/big/trying/tryingbigdata/controller/response/SearchResponse.java
@@ -1,6 +1,6 @@
 package com.data.big.trying.tryingbigdata.controller.response;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,6 +9,6 @@ import lombok.Getter;
 @Getter
 public class SearchResponse {
 
-    private Map<String, BigInteger> result;
+    private Map<String, BigDecimal> result;
 
 }
